+++ app/controllers/subjects_controller.rb
@@ -1,11 +1,11 @@
 class SubjectsController < ApplicationController
   before_action :authenticate_user!
   before_action :set_viewable_project, only: [ :index, :show, :timeline, :comments, :settings, :files, :events, :sheets, :event, :report ]
-  before_action :set_editable_project_or_editable_site, only: [ :new, :edit, :create, :update, :destroy, :search, :choose_site, :choose_date, :choose_an_event_for_subject, :data_entry, :choose_event, :launch_subject_event, :edit_event, :update_event ]
-  before_action :redirect_without_project, only: [ :index, :show, :timeline, :comments, :settings, :files, :sheets, :event, :report, :new, :edit, :create, :update, :destroy, :search, :choose_site, :choose_date, :choose_an_event_for_subject, :data_entry, :choose_event, :events, :launch_subject_event, :edit_event, :update_event ]
+  before_action :set_editable_project_or_editable_site, only: [ :new, :edit, :create, :update, :destroy, :search, :choose_site, :choose_date, :choose_an_event_for_subject, :data_entry, :choose_event, :launch_subject_event, :edit_event, :update_event, :destroy_event ]
+  before_action :redirect_without_project, only: [ :index, :show, :timeline, :comments, :settings, :files, :sheets, :event, :report, :new, :edit, :create, :update, :destroy, :search, :choose_site, :choose_date, :choose_an_event_for_subject, :data_entry, :choose_event, :events, :launch_subject_event, :edit_event, :update_event, :destroy_event ]
   before_action :set_viewable_subject, only: [ :show, :timeline, :comments, :settings, :files, :sheets, :event ]
-  before_action :set_editable_subject, only: [ :edit, :update, :destroy, :choose_date, :choose_an_event_for_subject, :data_entry, :choose_event, :events, :launch_subject_event, :edit_event, :update_event ]
-  before_action :redirect_without_subject, only: [ :show, :timeline, :comments, :settings, :files, :sheets, :event, :edit, :update, :destroy, :choose_date, :choose_an_event_for_subject, :data_entry, :choose_event, :events, :launch_subject_event, :edit_event, :update_event ]
+  before_action :set_editable_subject, only: [ :edit, :update, :destroy, :choose_date, :choose_an_event_for_subject, :data_entry, :choose_event, :events, :launch_subject_event, :edit_event, :update_event, :destroy_event ]
+  before_action :redirect_without_subject, only: [ :show, :timeline, :comments, :settings, :files, :sheets, :event, :edit, :update, :destroy, :choose_date, :choose_an_event_for_subject, :data_entry, :choose_event, :events, :launch_subject_event, :edit_event, :update_event, :destroy_event ]
 
   def data_entry
   end
@@ -36,7 +36,16 @@ class SubjectsController < ApplicationController
     else
       redirect_to [@project, @subject], alert: "#{params[:new_event_date].inspect}"
     end
+  end
 
+  def destroy_event
+    @event = @project.events.find_by_param(params[:event_id])
+    @subject_event = @subject.subject_events.where(event_id: @event.id).find_by_id(params[:subject_event_id]) if @event
+    if @subject_event
+      @subject_event.unlink_sheets!(current_user, request.remote_ip)
+      @subject_event.destroy
+    end
+    redirect_to [@project, @subject]
   end
 
   def events