+++ app/models/subject_event.rb
@@ -18,4 +18,10 @@ class SubjectEvent < ActiveRecord::Base
     self.event_date ? self.event_date.strftime("%a, %B %-d, %Y") : 'No Date'
   end
 
+  def unlink_sheets!(current_user, remote_ip)
+    self.sheets.each do |sheet|
+      SheetTransaction.save_sheet!(sheet, { subject_event_id: nil, last_user_id: current_user.id, last_edited_at: Time.now }, { }, current_user, remote_ip, 'sheet_update')
+    end
+  end
+
 end