+++ config/routes.rb
@@ -147,6 +147,7 @@ Rails.application.routes.draw do
         get "events/:event_id/:subject_event_id/:event_date", action: :event, as: :event
         get "events/:event_id/:subject_event_id/:event_date/edit", action: :edit_event, as: :edit_event
         post "events/:event_id/:subject_event_id/:event_date", action: :update_event, as: :update_event
+        delete "events/:event_id/:subject_event_id/:event_date", action: :destroy_event, as: :destroy_event
         get :timeline
         get :comments
         get :settings